+++ mininterface/_lib/dataclass_creation.py
@@ -1,11 +1,13 @@
 import re
 import warnings
-from dataclasses import MISSING, dataclass, fields, is_dataclass
+from dataclasses import MISSING, asdict, dataclass, fields, is_dataclass
 from types import UnionType
-from typing import Annotated, Optional, Type, Union, get_args, get_origin, TypeVar
+from typing import Annotated, Optional, Sequence, Type, Union, get_args, get_origin, TypeVar
+
 
 try:
     from tyro._singleton import MISSING_NONPROP
+    from tyro.extras import subcommand_type_from_defaults
 
     from ..cli import SubcommandPlaceholder
 except ImportError:
@@ -34,12 +36,6 @@ except ImportError:
     attr = None
 
 T = TypeVar("T")
-
-@dataclass(slots=True)
-class ChosenSubcommand:
-    name: str
-    subdict: dict
-
 
 def coerce_type_to_annotation(value, annotation):
     """
@@ -108,7 +104,14 @@ def _unwrap_annotated(tp):
     return tp
 
 
-def create_with_missing(env: T, disk: dict, wf: Optional[dict] = None, m: Optional["Mininterface"] = None)->T:
+def create_with_missing(
+    env: T,
+    disk: dict,
+    wf: Optional[dict] = None,
+    m: Optional["Mininterface"] = None,
+    subc: Optional[dict] = None,
+    subc_passage: Optional[list] = None,
+) -> T:
     """
     Create a default instance of an Env object. This is due to provent tyro to spawn warnings about missing fields.
     Nested dataclasses have to be properly initialized. YAML gave them as dicts only.
@@ -118,6 +121,8 @@ def create_with_missing(env: T, disk: dict, wf: Optional[dict] = None, m: Option
     And such fields are put into the wf (wrong_fields) dict.
 
     Having the `m` defined means we build the dataclass. None means we are still in the config file parsing context.
+
+    Can set `__subcommand_dialog_used` attribute to the `m`.
     """
     # NOTE a test is missing
     # @dataclass
@@ -148,7 +153,7 @@ def create_with_missing(env: T, disk: dict, wf: Optional[dict] = None, m: Option
     # as the default value takes the precedence over the hard coded one, even if missing.
     out = {}
     missings: list[Tag] = []
-    for name, v in proc_method(env, disk, wf, m):
+    for name, v in proc_method(env, disk, wf, m, subc, subc_passage):
         out[name] = v
         if v == MISSING_NONPROP and wf is not None:
             # For building config file, the MISSING_NONPROP is alright as we expect tyro to fail
@@ -168,28 +173,48 @@ def create_with_missing(env: T, disk: dict, wf: Optional[dict] = None, m: Option
     return model
 
 
-def _process_pydantic(env, disk, wf: Optional[dict], m: Optional["Mininterface"] = None):
+def _process_pydantic(
+    env,
+    disk,
+    wf: Optional[dict],
+    m: Optional["Mininterface"] = None,
+    subc: Optional[dict] = None,
+    subc_passage: Optional[list] = None,
+):
     for name, f in env.model_fields.items():
         if name in disk:
             default_value = f.default if f.default is not None else MISSING
-            v = _process_field(name, f.annotation, disk[name], wf, m, default_value)
+            v = _process_field(
+                name, f.annotation, disk[name], wf, m, default_value, subc=subc, subc_passage=subc_passage
+            )
         elif f.default is not None:
             v = f.default
         else:
-            v = _process_field(name, f.annotation, MISSING_NONPROP, wf, m, default_value=MISSING)
+            v = _process_field(
+                name, f.annotation, MISSING_NONPROP, wf, m, default_value=MISSING, subc=subc, subc_passage=subc_passage
+            )
         yield name, v
 
 
-def _process_attr(env, disk, wf: Optional[dict], m: Optional["Mininterface"] = None):
+def _process_attr(
+    env,
+    disk,
+    wf: Optional[dict],
+    m: Optional["Mininterface"] = None,
+    subc: Optional[dict] = None,
+    subc_passage: Optional[list] = None,
+):
     for f in attr.fields(env):
         has_default = f.default is not attr.NOTHING
         default_val = f.default if has_default else MISSING
         if f.name in disk:
-            v = _process_field(f.name, f.type, disk[f.name], wf, m, default_val)
+            v = _process_field(f.name, f.type, disk[f.name], wf, m, default_val, subc=subc, subc_passage=subc_passage)
         elif has_default:
             v = f.default
         else:
-            v = _process_field(f.name, f.type, MISSING_NONPROP, wf, m, default_val)
+            v = _process_field(
+                f.name, f.type, MISSING_NONPROP, wf, m, default_val, subc=subc, subc_passage=subc_passage
+            )
         yield f.name, v
 
 
@@ -219,82 +244,131 @@ def _resolve_ftype(ftype, default_value):
     return ftype
 
 
-def _init_struct_value(ftype, disk_value, wf, fname, m):
+def _init_struct_value(ftype, disk_value, wf, fname, m, subc, subc_passage, subsubc=None):
     """Init structural type (dataclass/attrs/pydantic)."""
     if wf is not None:
         subwf = wf[fname] = {}
     else:
         subwf = None
-    v = create_with_missing(ftype, disk_value, subwf, m)
+
+    if not subsubc:
+        if subc is not None:
+            subsubc = subc.get(fname, {})
+        else:
+            subsubc = None
+
+    v = create_with_missing(ftype, disk_value, subwf, m, subsubc, subc_passage)
     if wf and not subwf:  # prevent having an empty section to edit
         del wf[fname]
     return v
 
-
-def _process_field(fname, ftype, disk_value, wf, m, default_value=MISSING):
-    ftype = _resolve_ftype(ftype, default_value)
+def _is_subcommands(ftype):
     origin = _get_origin(ftype)
+    return (origin is Union or origin is UnionType) and all(_is_struct_type(cl) for cl in get_args(ftype))
 
-    # the subcommand has been already chosen by CLI parser
-    if isinstance(disk_value, ChosenSubcommand):  # `(class Message | class Console)`
-        # there might be a subcommand prefix, ex. 'val:message' -> 'message'
-        name = disk_value.name.partition(":")[2] or disk_value.name
-
-        for _subcomm in get_args(_unwrap_annotated(ftype)):
-            if name == to_kebab_case(_subcomm.__name__):
-                ftype = _subcomm  # `class Message` only
-                disk_value = disk_value.subdict
-                break
-        else:
-            raise ValueError(f"Type {disk_value} not found in {ftype}")
-
-    if disk_value is not MISSING_NONPROP:
-        if _is_struct_type(ftype):
-            return _init_struct_value(ftype, disk_value, wf, fname, m)
-        # NOTE for handling subcommands in config file, I'll need something like this,
-        # combined with a structure like ChosenSubcommand, possibly inheriting from MISSING_NONPROP
-        # to be ignored by tyro, yet fetchable by cli_parser
-        # elif _is_subcommands:
-        #     return {cl.__name__: _init_struct_value(cl, {}, wf, fname, m) for cl in get_args(ftype)}
-        return coerce_type_to_annotation(disk_value, ftype)
+def _process_field(
+    fname,
+    ftype,
+    disk_value,
+    wf,
+    m,
+    default_value=MISSING,
+    subc: Optional[dict] = None,
+    subc_passage: Optional[list] = None,
+):
+    ftype = _resolve_ftype(ftype, default_value)
 
     if _is_struct_type(ftype):
         # Ex. `foo: Subcommand`
-        return _init_struct_value(ftype, {}, wf, fname, m)
+        return _init_struct_value(
+            ftype, disk_value if disk_value is not MISSING_NONPROP else {}, wf, fname, m, subc, subc_passage
+        )
+    elif _is_subcommands(ftype):
+        # We must handle the case when there are multiple subcommands possible.
+        # The user decides now which way to go (choose a subcommand).
+        # Ex. `foo: Subcommand1 | Subcommand2`
 
-    # We must handle the case when there are multiple subcommands possible.
-    # The user decides now which way to go (choose a subcommand).
-    # Ex. `foo: Subcommand1 | Subcommand2`
-    _is_subcommands = (origin is Union or origin is UnionType) and all(_is_struct_type(cl) for cl in get_args(ftype))
-    if _is_subcommands:
-        if not m:  # we are parsing the config file. The fields are not defined in the config file
+        env_classes = get_args(ftype)
+        if subc_passage is None and disk_value is not MISSING_NONPROP:
+            # We are parsing the config file and the fields are defined in the config file.
+            if subc is not None:
+                subc[fname] = {
+                    to_kebab_case(cl.__name__): disk_value.get(to_kebab_case(cl.__name__), {}) for cl in env_classes
+                }
+                return MISSING_NONPROP
+            else:
+                raise ValueError("Unknown subcommand config parsing")
+
+        if subc_passage:
+            # the next subcommand has been already chosen by CLI parser
+            ftype, _ = pop_from_passage(subc_passage, env_classes)
+        elif not m:
+            # We are parsing the config file and the fields are not defined in the config file.
+            # Or we're dealing with the --help while `subc_passage` did not crawled till the end.
+            #
+            # Ex. `./program.py c:env1 --help` (NOT till the end)
+            # Ex. `./program.py c:env1 c.c2:subc1 --help` (till the end)
             return MISSING_NONPROP
-        ftype = choose_subcommand(get_args(ftype), m)
-        return _init_struct_value(ftype, {}, wf, fname, m)
+        else:
+            ftype = choose_subcommand(env_classes, m)
+            setattr(m, "__subcommand_dialog_used", True)
 
-    return MISSING_NONPROP
+        # unwrap the subc, choosing the chosen class for the default value
+        disk_val = v if subc and (v := subc.get(fname, {}).get(to_kebab_case(ftype.__name__))) else {}
+        return _init_struct_value(ftype, disk_val, wf, fname, m, None, subc_passage, subsubc=disk_val)
+
+    if disk_value is not MISSING_NONPROP:
+        return coerce_type_to_annotation(disk_value, ftype)
+    else:
+        return MISSING_NONPROP
 
 
-def _process_dataclass(env, disk, wf: Optional[dict], m: Optional["Mininterface"] = None):
+def _process_dataclass(
+    env,
+    disk,
+    wf: Optional[dict],
+    m: Optional["Mininterface"] = None,
+    subc: Optional[dict] = None,
+    subc_passage: Optional[list] = None,
+):
     for f in fields(_unwrap_annotated(env)):
 
         if f.name.startswith("__"):
             continue
         elif f.name in disk:
-            v = _process_field(f.name, f.type, disk[f.name], wf, m, f.default)
+            v = _process_field(f.name, f.type, disk[f.name], wf, m, f.default, subc=subc, subc_passage=subc_passage)
         elif f.default_factory is not MISSING:
             v = f.default_factory()
         elif f.default is not MISSING:
             v = f.default
         else:
-            v = _process_field(f.name, f.type, MISSING_NONPROP, wf, m)
+            v = _process_field(f.name, f.type, MISSING_NONPROP, wf, m, subc=subc, subc_passage=subc_passage)
         yield f.name, v
 
 
 def choose_subcommand(env_classes: list[Type[DataClass]], m: "Mininterface[EnvClass]"):
     # NOTE make select display buttons if there is a little amount of options.
-    env = m.select({(to_kebab_case(cl.__name__).replace("-", " ").capitalize(), _get_parser(cl).description): cl for cl in env_classes  if cl is not SubcommandPlaceholder})
+    env = m.select(
+        {
+            (to_kebab_case(cl.__name__).replace("-", " ").capitalize(), _get_parser(cl).description): cl
+            for cl in env_classes
+            if cl is not SubcommandPlaceholder
+        }
+    )
     return env
+
+def pop_from_passage(passage, env_classes: Sequence[T])-> tuple[T, str]:
+    cl_name = passage.pop(0)
+    # there might be a subcommand prefix, ex. 'val:message' -> 'message'
+    cl_name = cl_name.partition(":")[2] or cl_name
+    ftype = get_chosen(cl_name, env_classes)
+    return ftype, cl_name
+
+def get_chosen(cl_name, env_classes):
+    for cl in env_classes:
+        if to_kebab_case(cl.__name__) == cl_name:
+            return cl
+    raise ValueError(f"Type {cl_name} not found in {env_classes}")
 
 def to_kebab_case(name: str) -> str:
     """MyClass -> my-class"""